Added enemy power ups, added death state, added collision for enemy power ups, added allmighty candy

+++ GameAI/pathfinding/game/PacSteering.h
@@ -30,7 +30,7 @@ private:
 	Path mPath;
 	//Player mPlayerSteering;
 	bool isUnitDown, isUnitRight, isUnitLeft, isUnitUp;
-	int mSpeed, mSpeedX, mSpeedY;
+	float mSpeed, mSpeedX, mSpeedY;
 	int mIndex;
 
 	string oldDir, newDir;
